+++ changelog.md
@@ -11,7 +11,7 @@ version 5.1.0 -- upcoming
   * added result environment to test plugin
   * fixed bugs in test-plugin that lead to re-execution of tests
   * added support for line-breaks in `x-www-form-urlencoded` mimetype requests
-  * 
+  * improved dirty-indicator for request tabs
 
 version 5.0.0
   * included [cassandra plugin](/milkman-cassandra) and [Grpc plugin](/milkman-grpc) into macos app bundle